# jm-ai-action-hub-focus-suite-v0.9.0-ios-v0.2.1/contracts/action-hub.openapi-v0.9.0.json
{
  "components": {
    "schemas": {
...
      "ApprovalRequest": {
        "properties": {
          "actor": {
            "default": "user",
            "title": "Actor",
            "type": "string"
          },
          "force_review_items": {
            "default": false,
            "title": "Force Review Items",
            "type": "boolean"
          },
          "item_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Item Ids"
          }
        },
        "title": "ApprovalRequest",
        "type": "object"
      },
...
      "ConnectorStatus": {
        "properties": {
          "configured": {
            "title": "Configured",
            "type": "boolean"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "execution_mode": {
            "title": "Execution Mode",
            "type": "string"
          },
          "healthy": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Healthy"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "configured",
          "execution_mode",
          "detail"
        ],
        "title": "ConnectorStatus",
        "type": "object"
      },
      "DailyBrief": {
        "properties": {
          "ai_ready": {
            "items": {
              "$ref": "#/components/schemas/BriefItem"
            },
            "title": "Ai Ready",
            "type": "array"
          },
          "date": {
            "title": "Date",
            "type": "string"
          },
          "due_tasks": {
            "items": {
              "$ref": "#/components/schemas/BriefItem"
            },
            "title": "Due Tasks",
            "type": "array"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/BriefItem"
            },
            "title": "Events",
            "type": "array"
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "needs_review": {
            "items": {
              "$ref": "#/components/schemas/BriefItem"
            },
            "title": "Needs Review",
            "type": "array"
          },
          "overdue": {
            "items": {
              "$ref": "#/components/schemas/BriefItem"
            },
            "title": "Overdue",
            "type": "array"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          },
          "waiting": {
            "items": {
              "$ref": "#/components/schemas/BriefItem"
            },
            "title": "Waiting",
            "type": "array"
          }
        },
        "required": [
          "generated_at",
          "timezone",
          "date",
          "events",
          "due_tasks",
          "overdue",
          "needs_review",
          "summary"
        ],
        "title": "DailyBrief",
        "type": "object"
      },
...
      "ExecuteRequest": {
        "properties": {
          "actor": {
            "default": "user",
            "title": "Actor",
            "type": "string"
          },
          "drain_inline": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Drain Inline"
          },
          "item_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Item Ids"
          },
          "retry_failed": {
            "default": false,
            "title": "Retry Failed",
            "type": "boolean"
          }
        },
        "title": "ExecuteRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "InboxParseRequest": {
        "properties": {
          "force_new": {
            "default": false,
            "title": "Force New",
            "type": "boolean"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "reference_time": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference Time"
          },
          "source": {
            "default": "paste",
            "maxLength": 40,
            "title": "Source",
            "type": "string"
          },
          "text": {
            "minLength": 1,
            "title": "Text",
            "type": "string"
          },
          "timezone": {
            "default": "Asia/Seoul",
            "title": "Timezone",
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "title": "InboxParseRequest",
        "type": "object"
      },
...
      "PlanRead": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "inbox": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InboxRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "inbox_id": {
            "title": "Inbox Id",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ActionItemRead"
            },
            "title": "Items",
            "type": "array"
          },
          "parser_name": {
            "title": "Parser Name",
            "type": "string"
          },
          "reference_time": {
            "format": "date-time",
            "title": "Reference Time",
            "type": "string"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/PlanStatus"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "inbox_id",
          "status",
          "parser_name",
          "summary",
          "reference_time",
          "revision",
          "created_at",
          "updated_at",
          "items"
        ],
        "title": "PlanRead",
        "type": "object"
      },
      "PlanStatus": {
        "enum": [
          "draft",
          "approved",
          "queued",
          "routed",
          "partial",
          "completed",
          "failed",
          "rejected"
        ],
        "title": "PlanStatus",
        "type": "string"
      },
...
      "RejectRequest": {
        "properties": {
          "actor": {
            "default": "user",
            "title": "Actor",
            "type": "string"
          },
          "item_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Item Ids"
          },
          "reason": {
            "default": "",
            "title": "Reason",
            "type": "string"
          }
        },
        "title": "RejectRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "ActionHubAdminKey": {
        "description": "Administrative Action Hub API key. Never embed this key in the iOS app.",
        "in": "header",
        "name": "X-Action-Hub-Key",
        "type": "apiKey"
      },
      "ActionHubMobileBearer": {
        "description": "Short-lived mobile access token issued after QR pairing.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Approval-based closed-loop action control for people, AI workers, and external systems.",
    "title": "JM-AI Action Hub",
    "version": "0.9.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/audit": {
      "get": {
        "operationId": "audit_log_api_v1_audit_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "entity_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Entity Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response Audit Log Api V1 Audit Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "ActionHubAdminKey": []
          }
        ],
        "summary": "Audit Log"
      }
    },
    "/api/v1/brief/today": {
      "get": {
        "operationId": "today_brief_api_v1_brief_today_get",
        "parameters": [
          {
            "in": "query",
            "name": "at",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "At"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DailyBrief"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "ActionHubAdminKey": []
          }
        ],
        "summary": "Today Brief"
      }
    },
    "/api/v1/connectors/status": {
      "get": {
        "operationId": "connector_status_api_v1_connectors_status_get",
        "parameters": [
          {
            "in": "query",
            "name": "probe",
            "required": false,
            "schema": {
              "default": false,
              "title": "Probe",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ConnectorStatus"
                  },
                  "title": "Response Connector Status Api V1 Connectors Status Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "ActionHubAdminKey": []
          }
        ],
        "summary": "Connector Status"
      }
    },
...
    "/api/v1/exports/{filename}": {
      "get": {
        "operationId": "download_export_api_v1_exports__filename__get",
        "parameters": [
          {
            "in": "path",
            "name": "filename",
            "required": true,
            "schema": {
              "title": "Filename",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "ActionHubAdminKey": []
          }
        ],
        "summary": "Download Export"
      }
    },
...
    "/api/v1/inbox/parse": {
      "post": {
        "operationId": "parse_inbox_api_v1_inbox_parse_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InboxParseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "ActionHubAdminKey": []
          }
        ],
        "summary": "Parse Inbox"
      }
    },
...
    "/api/v1/plans": {
      "get": {
        "operationId": "list_plans_api_v1_plans_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PlanRead"
                  },
                  "title": "Response List Plans Api V1 Plans Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "ActionHubAdminKey": []
          }
        ],
        "summary": "List Plans"
      }
    },
    "/api/v1/plans/{plan_id}": {
      "get": {
        "operationId": "read_plan_api_v1_plans__plan_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "plan_id",
            "required": true,
            "schema": {
              "title": "Plan Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "ActionHubAdminKey": []
          }
        ],
        "summary": "Read Plan"
      }
    },
    "/api/v1/plans/{plan_id}/approve": {
      "post": {
        "operationId": "approve_api_v1_plans__plan_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "plan_id",
            "required": true,
            "schema": {
              "title": "Plan Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "ActionHubAdminKey": []
          }
        ],
        "summary": "Approve"
      }
    },
    "/api/v1/plans/{plan_id}/execute": {
      "post": {
        "operationId": "execute_api_v1_plans__plan_id__execute_post",
        "parameters": [
          {
            "in": "path",
            "name": "plan_id",
            "required": true,
            "schema": {
              "title": "Plan Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecuteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "ActionHubAdminKey": []
          }
        ],
        "summary": "Execute"
      }
    },
    "/api/v1/plans/{plan_id}/items/{item_id}": {
      "patch": {
        "operationId": "patch_item_api_v1_plans__plan_id__items__item_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "plan_id",
            "required": true,
            "schema": {
              "title": "Plan Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "title": "Item Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-actor",
            "required": false,
            "schema": {
              "default": "user",
              "title": "X-Actor",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionItemUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "ActionHubAdminKey": []
          }
        ],
        "summary": "Patch Item"
      }
    },
    "/api/v1/plans/{plan_id}/reject": {
      "post": {
        "operationId": "reject_api_v1_plans__plan_id__reject_post",
        "parameters": [
          {
            "in": "path",
            "name": "plan_id",
            "required": true,
            "schema": {
              "title": "Plan Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "ActionHubAdminKey": []
          }
        ],
        "summary": "Reject"
      }
    },
...
  }
}